requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
attrs==19.3.0             # via jsonschema
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via mozilla-schema-generator
dj-database-url==0.5.0    # via -r requirements.in
django-configurations==2.2  # via -r requirements.in
django==3.0.8             # via -r requirements.in
google-api-core==1.21.0   # via google-cloud-bigquery, google-cloud-core
google-auth==1.18.0       # via google-api-core, google-cloud-bigquery
google-cloud-bigquery==1.24.0  # via -r requirements.in
google-cloud-core==1.3.0  # via google-cloud-bigquery
google-resumable-media==0.5.1  # via google-cloud-bigquery
googleapis-common-protos==1.52.0  # via google-api-core
idna==2.10                # via requests
jsonschema==3.2.0         # via mozilla-schema-generator
mozilla-schema-generator==0.1.4  # via -r requirements.in
protobuf==3.12.2          # via google-api-core, google-cloud-bigquery, googleapis-common-protos
psycopg2-binary==2.8.5    # via -r requirements.in
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pyrsistent==0.16.0        # via jsonschema
pytz==2020.1              # via django, google-api-core
pyyaml==5.3.1             # via mozilla-schema-generator
requests==2.24.0          # via google-api-core, mozilla-schema-generator
rsa==4.6                  # via google-auth
six==1.15.0               # via django-configurations, google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, jsonschema, protobuf, pyrsistent
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
